Add files via upload
This project focuses on sentiment analysis of Amazon product reviews using two different techniques: VADER (Valence Aware Dictionary and sEntiment Reasoner) and a pre-trained RoBERTa model. Sentiment analysis is a crucial task in natural language processing (NLP) that helps in understanding customer opinions and feedback.

Key Features:
Dataset Exploration: The project begins with an exploration of the dataset, including the distribution of review scores and basic statistics.
Sentiment Analysis with VADER: VADER, a lexicon-based sentiment analysis tool, is employed to assign sentiment scores (negative, neutral, positive, and compound) to each review.
Sentiment Analysis with RoBERTa: A pre-trained RoBERTa model, fine-tuned for sentiment analysis, is used to obtain sentiment scores (negative, neutral, and positive) for each review.
Comparison of Results: The results from VADER and RoBERTa are compared and visualized to understand the effectiveness of each method in capturing sentiment.
Example Reviews: Examples of reviews from different sentiment categories are provided to illustrate the performance of both VADER and RoBERTa in analyzing sentiment.
Conclusion:
This project demonstrates two different approaches to sentiment analysis and provides insights into the strengths and weaknesses of each method. It serves as a resource for individuals interested in sentiment analysis, NLP, and machine learning techniques applied to text data.